feature(pass-data): pass data to chat component through
